Split kernel host capability modules
